update stop rebalancing test

+++ programs/pie/src/instructions/mint_basket_token.rs
@@ -60,12 +60,12 @@ pub fn mint_basket_token(ctx: Context<MintBasketTokenContext>, amount: u64) -> R
             .iter()
             .find(|a| a.mint == token_config.mint)
         {
-            let possible_mint = user_asset
-                .amount
-                .checked_mul(EXPONENT)
+            let possible_mint = (user_asset
+                .amount as u128)
+                .checked_mul(EXPONENT as u128)
                 .unwrap()
-                .checked_div(token_config.ratio)
-                .unwrap();
+                .checked_div(token_config.ratio as u128)
+                .unwrap() as u64;
             amount_can_mint = amount_can_mint.min(possible_mint);
         }
     }
@@ -78,14 +78,14 @@ pub fn mint_basket_token(ctx: Context<MintBasketTokenContext>, amount: u64) -> R
             .iter_mut()
             .find(|a| a.mint == token_config.mint)
         {
-            let amount_to_deduct = token_config.ratio
-                .checked_mul(amount)
+            let amount_to_deduct = (token_config.ratio as u128)
+                .checked_mul(amount as u128)
                 .unwrap()
-                .checked_div(EXPONENT)
+                .checked_div(EXPONENT as u128)
                 .ok_or(PieError::InsufficientBalance)?;
             asset.amount = asset
                 .amount
-                .checked_sub(amount_to_deduct)
+                .checked_sub(amount_to_deduct as u64)
                 .ok_or(PieError::InsufficientBalance)?;
         }
     }
